almazny: merged with mirny removed few ports (mirny cpld limitation)
